Restructure discogs service files
